requirements.txt
aiogram==3.7.0
google-genai
Pillow
python-dotenv
